refactor upscaling scoring methods
